requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
axial_positional_embedding==0.3.12
braindecode==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
CoLT5-attention==0.11.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deepspeed==0.18.0
docstring-inheritance==3.0.0
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.1
h5py==3.14.0
hf-xet==1.1.10
hjson==3.1.0
huggingface-hub==0.35.3
hyper-connections==0.2.1
idna==3.10
jieba==0.42.1
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
linear-attention-transformer==0.19.1
linformer==0.2.3
local-attention==1.11.2
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mne==1.10.1
mne-bids==0.17.0
mpi4py 
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
product_key_memory==0.3.0
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
pypinyin==0.55.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
rouge_score==0.1.2
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.41.0
six==1.17.0
skorch==1.3.1
smmap==5.0.2
soundfile==0.13.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
textdistance==4.6.3
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.22.1
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.57.0
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.1
Werkzeug==3.1.3
wfdb==4.3.0
yarl==1.22.0
zhconv==1.4.3
